examples/cubesandbox-base-nginx/requirements.txt
e2b>=2.4.1
python-dotenv
